glibc: Rework locale-gen and improve stripping.
I didn't think to include the STRIP_* variables defined in
/etc/makepkg.conf which resulted in many missed opportunities and thus a
larger resulting package.  Correctly stripping shared, static and
binaries saved a futher 8M.

Locale-gen was also changed to be a bit simpler.  Error out if the
locales file is empty now.  I had intended on allowing users to add
locales via locale-gen but decided against it as it merely replicates
redirection with added complexity which don't appear to be justified.