ncurses: provide tinfo and introduce hack for it
The tinfo library is now built as a seperate library in Arch Linux's
package.  This has resulted in many linking failures as several projects
don't use pkg-config or ncurses-config to determine the correct linking
flags.

The workaround introduced is to implicitly add -ltinfo using a loader
script when projects attempt to link against libncursesw.

Although this introduces two -ltinfo flags for projects correctly
linking against ncurses, it should be harmless.