system-config: use hooks to patch owned files
This is a proof of concept attempt to evaluate how using pacman hooks
and patch(1) can be used to modify files which are already owned by
other packages.

The idea is to trigger on a list of matching target files and pass them
to a script which will detect if a corresponding patch is available.

It then checks if the patch has already been applied (indicated by a
reverse application succeeding in a dry run) before patching the file.