linux-mainline: drop versioned extramodules
As I cannot guarantee API/ABI stability for out of tree modules there is
no need to provide a weak-modules directory.  Instead we leave it
unversioned which requires out of tree modules be rebuilt for every
update.

This also results in much simpler packaging for out of tree modules such
as acpi_call.