linux-mainline: big changes
The main visible change is the addition of the generate-sundry script,
this is used to create the various hooks and preset files based on given
arguments.

This, and combined with various changes throughout the PKGBUILD, allows
me to generalise the entire thing over the pkgname.  That is, we can now
use pkgname=linux or pkgname=linux-foo and both should generate a
respectable package.

In the case of pkgname=linux there simply won't be any -variant tags,
such as -ARCH.  Directories like /lib/modules/extramodules-ARCH will
simply become /lib/modules/extramodules.

Some minor changes is the removal of depmod wrangling.  It turns out
that the depmod.sh script was fixed to use the result of INSTALL_MOD_PATH
back in 2011 so we can quite happily let modules_install run the script
internally.

Finally I've rearranged the package() function a bit so that each
subsection is more obvious.  I.e. the kernel and module installation,
the api headers, the internal headers, the documentation, and the clean
up.